+++ frontend/src/pages/DataCatalog/MethodologyTab.tsx
@@ -1,18 +1,12 @@
 import Grid from '@mui/material/Grid'
 import styles from './DataCatalogPage.module.scss'
-import {
-  CONTACT_TAB_LINK,
-  HET_URL,
-  DATA_TAB_LINK,
-} from '../../utils/internalRoutes'
+import { CONTACT_TAB_LINK, DATA_TAB_LINK } from '../../utils/internalRoutes'
 import { Helmet } from 'react-helmet-async'
 import { getHtml, LinkWithStickyParams } from '../../utils/urlutils'
-import { selectFaqs } from '../WhatIsHealthEquity/FaqTab'
 import { METRIC_CONFIG } from '../../data/config/MetricConfig'
 import { Card } from '@mui/material'
 import { urlMap } from '../../utils/externalUrls'
 import DefinitionsList from '../../reports/ui/DefinitionsList'
-import { currentYear } from '../../Footer'
 import {
   ALASKA_PRIVATE_JAIL_CAVEAT,
   CombinedIncarcerationStateMessage,
@@ -29,8 +23,8 @@ import {
 } from './methodologyContent/missingDataBlurbs'
 import { SHOW_PHRMA } from '../../data/providers/PhrmaProvider'
 import { HashLink } from 'react-router-hash-link'
-
-export const CITATION_APA = `Health Equity Tracker. (${currentYear()}). Satcher Health Leadership Institute. Morehouse School of Medicine. ${HET_URL}.`
+import { CITATION_APA } from '../../cards/ui/Sources'
+import { selectFaqs } from '../WhatIsHealthEquity/FaqData'
 
 function MethodologyTab() {
   return (
